% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agebiasData.R
\name{agebiasData}
\alias{agebiasData}
\title{Simulates age:bias data}
\usage{
agebiasData(ages, bias = c("under", "over", "unbiased"), seed)
}
\arguments{
\item{ages}{= simulated age data, bias = type of bias ("under","over","unbiased")}
}
\value{
Age data
}
\description{
Simulates biased age data from the simulated age sample
}
