% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthbins.R
\name{lengthbins}
\alias{lengthbins}
\title{Assigns individual length data to a length bin}
\usage{
lengthbins(len, by, dir = c("up", "down"))
}
\arguments{
\item{len}{= length data}

\item{by}{= sets the width of the length bin}

\item{dir}{= round "up" or round "down"}
}
\value{
length bin
}
\description{
Assigns individual length data to a length bin
}
