% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simalkData.R
\name{simalkData}
\alias{simalkData}
\title{Individual-based model to simulate age:length data for aged and unaged fish using a von Bertalanffy growth function}
\usage{
simalkData(ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv, seed)
}
\arguments{
\item{ages}{= ages produced by simageData}

\item{Linf.mu}{= Linf mean}

\item{Linf.cv}{= CV of Linf}

\item{k.mu}{= k mean}

\item{k.cv}{= CV of k}

\item{t0.mu}{= t0 mean}

\item{t0.cv}{= CV of t0}

\item{seed}{= seed}
}
\value{
Simulated length data
}
\description{
Simulates length and age data for individual fish
}
