% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simElapsed.R
\name{simElapsed}
\alias{simElapsed}
\title{Individual-based model to simulate mean length at annulus for the ith fish of age a}
\usage{
simElapsed(ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv, seed)
}
\arguments{
\item{ages}{= ages simulated from simageData function}

\item{Linf.mu}{= Linf mean}

\item{Linf.cv}{= CV of Linf}

\item{k.mu}{= k mean}

\item{k.cv}{= CV of k}

\item{t0.mu}{= t0 mean}

\item{t0.cv}{= CV of t0}

\item{seed}{= seed}
}
\value{
Simulated length data
}
\description{
Simulates data to estimate back-calculated lengths of age a for the ith fish
}
