% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVBGFlength.R, R/simlengthVBGF.R
\name{simVBGFlength}
\alias{simVBGFlength}
\alias{simVBGFlength}
\title{Individual-based model to simulate age:length data from a von Bertalanffy growth function}
\usage{
simVBGFlength(ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv, seed)

simVBGFlength(ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv, seed)
}
\arguments{
\item{ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv}{= von Bert growth parameters, seed=seed}

\item{ages, Linf.mu, Linf.cv, k.mu, k.cv, t0.mu, t0.cv}{= von Bert growth parameters, seed=seed}
}
\value{
Simulated length data

Simulated length data
}
\description{
Simulates length data for individual i of age a

Simulates length data for individual i of age a
}
